# swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Tienda y Pasarela de Pago (TPP)
  description: Esta es la documentación para la API del Microservicio de Tienda y
    Pasarela de Pago (TPP) del proyecto OverSounds.
  version: 2.0.0
...
servers:
- url: http://localhost
...
paths:
...
  /payment:
    get:
      tags:
      - payment
      summary: Returns a list of payment methods for the selected user.
      description: Returns a list of payment methods for the selected user.
      operationId: show_user_payment_methods
      responses:
        "200":
          description: Payment methods returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
                x-content-type: application/json
        "400":
          description: Bad request.
        "500":
          description: Generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - oversound_auth:
        - read:payment
      x-openapi-router-controller: swagger_server.controllers.payment_controller
    post:
      tags:
      - payment
      summary: Add a new payment method
      description: Add a new payment method
      operationId: add_payment_method
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethod"
      responses:
        "200":
          description: Payment method successfully added.
        "400":
          description: Invalid input.
        "500":
          description: Generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - oversound_auth:
        - write:payment
      x-openapi-router-controller: swagger_server.controllers.payment_controller
  /payment/{paymentMethodId}:
    delete:
      tags:
      - payment
      summary: Delete a payment method by ID.
      description: Delete a payment method by ID.
      operationId: delete_payment_method
      parameters:
      - name: paymentMethodId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string   
      responses:
        "200":
          description: Payment method successfully deleted.
        "400":
          description: Invalid ID supplied.
        "404":
          description: Payment method not found.
        "500":
          description: Generic error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - oversound_auth:
        - write:payment
        - read:payment
      x-openapi-router-controller: swagger_server.controllers.payment_controller
...
components:
  schemas:
...
    PaymentMethod:
      required:
      - cardHolder
      - cardNumber
      - expireMonth
      - expireYear
      type: object
      properties:
        cardNumber:
          type: string
          description: Masked card number.
          example: '**** **** **** 1234'
        expireMonth:
          maximum: 12
          minimum: 1
          type: integer
          description: Expire month of the card.
          example: 12
        expireYear:
          type: integer
          description: Expire year of the card.
          example: 2030
        cardHolder:
          type: string
          description: Name of the card
          example: John Doe
      description: A simple Payment Method object for storing card details.
      example:
        expireMonth: 12
        cardHolder: John Doe
        expireYear: 2030
        cardNumber: '**** **** **** 1234'
...
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
  securitySchemes:
    oversound_auth:
      type: apiKey
      name: token
      in: cookie
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_oversounds_auth